CsvToDb: Support replace_content for tables with ingoing references.
Until now, CsvToDb.replace_content could not be used for rows that are referenced by other tables because the former approach was to truncate the entire table before. This was not only slow but also would have broken existing references.

Now, the delta is computed manually so that only old rows that are no longer contained in the CSV file will be deleted from the database. This makes it possible for clients to have dangling rows removed from their tables efficiently. If a table is referenced by somewhere else, deletion is still possible by specifying "ON DELETE CASCADE" in the relevant constraint.

The new behavior is also tested.