Disable Instagram thumbnails temporarily
We're currently hitting rate limits (blocked IP) for instaloader because of too many requests due to recent pipeline tests. For now, let's turn off this data source for a couple of days.

In the long term, we should find a solution that does not produce this problem whenever we run the CI a couple of times at a single day.