cache spacy models across containers
Instead of downlaoding the spacy model again and again within each Dockerfile which was not only time- but also space-consuming, extract this operation into its own FetchSpacyModel task. This task is only invoked when the model is actually required. The task caches the model across containers and only updates the cache if a newer version is available.